requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
easyocr>=1.7.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
